Refactored things into multiple files with renaming and restructuring
