+++ crypto/src/main/scala/org/bitcoins/crypto/ECKey.scala
@@ -398,6 +398,10 @@ case class ECPublicKey(private val _bytes: ByteVector)
   def toXOnly: XOnlyPubKey = XOnlyPubKey(bytes.drop(1))
 
   def parity: KeyParity = KeyParity.fromByte(bytes.head)
+
+  def negate: ECPublicKey = {
+    ECPublicKey.fromBytes(parity.negate.bytes ++ bytes.drop(1))
+  }
 }
 
 object ECPublicKey extends Factory[ECPublicKey] {
