+++ crypto/src/main/scala/org/bitcoins/crypto/musig/MuSigTweak.scala
@@ -1,0 +1,7 @@
+package org.bitcoins.crypto.musig
+
+import org.bitcoins.crypto.{ECPublicKey, FieldElement}
+
+case class MuSigTweak(tweak: FieldElement, isXOnlyT: Boolean) {
+  def point: ECPublicKey = tweak.getPublicKey
+}
